step activity added

+++ app/src/main/java/com/example/android/baking/model/Ingredient.java
@@ -5,7 +5,7 @@ import android.os.Parcelable;
 
 public class Ingredient implements Parcelable {
 
-    private int quantity;
+    private double quantity;
     private String measure;
     private String ingredient;
 
@@ -14,14 +14,14 @@ public class Ingredient implements Parcelable {
     }
 
     protected Ingredient(Parcel in) {
-        quantity = in.readInt();
+        quantity = in.readDouble();
         measure = in.readString();
         ingredient = in.readString();
     }
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(quantity);
+        dest.writeDouble(quantity);
         dest.writeString(measure);
         dest.writeString(ingredient);
     }
@@ -43,11 +43,11 @@ public class Ingredient implements Parcelable {
         }
     };
 
-    public int getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(int aQuantity) {
+    public void setQuantity(double aQuantity) {
         quantity = aQuantity;
     }
